This PR contains a lot of refactors, mostly to do with adding support for arbitrary sample widths and arbitrary beamracing overlays on the framebuffer but the actual **features** addressed are:
- xbeam: Add an overlay cross/grid (selectable) in `xbeam`, adjust all time and voltage scales to include units and ensure the grid overlay matches the voltage calibration.
- xbeam: Add suffixes to a few options in `xbeam` that have physical meaning, also add `/d` or `d` suffix to all options that refer to 'divisions' e.g. volts/div or ms/div
- xbeam/all: Move resampling logic out of scope/vector RTL and up into top-level bitstreams, so it can be shared between scopes and vectorscopes. This increases the amount of upsampling in the quad-scope mode of `xbeam` by about 4x, as a result, faster timebases look good. So I added a `50us` timebase in addition to the existing ones.
- xbeam/selftest/all: Improve support for audio sample widths > 16 bits. Increase default bitwidth in `xbeam` and `selftest` audio pipeline, so both support up to approx +/- 10V (ADCs) and +/- 9V (DACs), whereas we were stuck with +/- 8.192V previously
- xbeam: swap contents of scope1 and scope2 pages. Scope2 now has the most useful options and is on the last page, which is easier to reach with the encoder. Also, add a `help` option to the `MISC` page (on by default). When off, the `help` page is gone, making it much easier to switch between vector and scope modes.
- scope/vector/all: add a proper rust HAL for the scope peripherals, to make them more easily reuseable.
- raster: split ASQ (audio) and PSQ (plotting) fixed-point types - previously everything was ASQ, but with this change, it's possible to build `xbeam` with a completely 24-bit audio path, while staying under 90% LUTs. Default audio path upgraded from 16-bits to 18-bits. ENOB is around there anyway. 
- calibration: fix some bugs in cal loading so that things should still work if a selftest with 'wide' ASQ writing a cal is still read correctly with user bitstreams assuming a 'not-wide' ASQ, and vice-versa.

+++ gateware/src/top/sampler/fw/src/channel.rs
@@ -17,7 +17,7 @@ impl<G: GrainPlayer> Channel<G> {
     }
 
     /// Update grain player from channel options and input state
-    pub fn update(&mut self, opts: &ChannelOpts, max_samples: u32, touch_idx: usize, touch: &[u8; 8], jack: u8, cv: i16) {
+    pub fn update(&mut self, opts: &ChannelOpts, max_samples: u32, touch_idx: usize, touch: &[u8; 8], jack: u8, cv: i32) {
         let size = max_samples;
         let ui_start = (opts.start.value as i32).max(5) as u32;
         let start = size.saturating_sub(ui_start);
